Fix publisher alertiing severity level
The check for which severity level for the alert for Publisher was
incorrect. Publisher does not have a deployment_environment therefore
the severity level would have always defaulted to low.

Correct this, so now alerts for Live will go to the correct Slack
channel.

+++ scripts/deploy_alerting.rb
@@ -14,26 +14,30 @@ out_path = './out.yml'
 raise ArgumentError.new('Please provide namespace/alerts') if alert.nil?
 raise ArgumentError.new('Please provide platform environment') if platform_env.nil?
 raise ArgumentError.new('Please provide deployment environment') if deployment_env.nil? && (ARGV[0] != 'publisher')
+raise ArgumentError.new('Publisher does not require a deployment environment') if ARGV[0] == 'publisher' && deployment_env
 
 FileUtils.rm(out_path, force: true)
 
 File.open(out_path, 'w') do |f|
-  env_string = "#{platform_env}-#{deployment_env}"
-  severity = 'form-builder' if env_string == 'live-production'
+  env_string = ARGV[0] == 'publisher' ? platform_env : "#{platform_env}-#{deployment_env}"
+  puts "Environment string => #{env_string}"
+
+  severity = 'form-builder' if %w(live-production live).include?(env_string)
+  puts "Severity => #{severity}"
+
   string = File.read(alert)
   template = ERB.new(string)
   f.write template.result(binding)
   f.write "---\n"
 end
 
+puts "Applying monitoring config for #{ARGV[0]}"
 `kubectl apply -f #{out_path}`
 
 if $?.success?
   puts "Successfully applied changes for #{alert} to #{platform_env} #{deployment_env}"
 else
-  puts "🔥🔥🔥🔥🔥🔥\n\n"
-  puts "Unable to apply changes for #{alert} to #{platform_env} #{deployment_env}"
-  puts "\n\n🔥🔥🔥🔥🔥🔥"
+  raise StandardError.new("Unable to apply changes for #{alert} to #{platform_env} #{deployment_env}")
 end
 
 FileUtils.rm(out_path, force: true)
